Separate error group callback tests
